+++ src/renderer/components/contact/Contact.tsx
@@ -1,7 +1,7 @@
 import React, { CSSProperties, PropsWithChildren } from 'react'
 import { C } from 'deltachat-node/dist/constants'
-import classNames from 'classnames'
 import { ContactJSON } from '../../../shared/shared-types'
+import { Avatar } from '../Avatar'
 
 export function convertContactProps(contact: ContactJSON) {
   return {
@@ -12,67 +12,6 @@ export function convertContactProps(contact: ContactJSON) {
     isMe: contact.id === C.DC_CONTACT_ID_SELF,
     verified: contact.isVerified,
   }
-}
-
-export function renderAvatar(
-  avatarPath: string,
-  color: string,
-  displayName: string
-) {
-  return (
-    <Avatar avatarPath={avatarPath} color={color} displayName={displayName} />
-  )
-}
-
-export function Avatar(props: {
-  avatarPath?: string
-  color?: string
-  displayName: string
-  large?: boolean
-}) {
-  const { avatarPath, color, displayName, large } = props
-  if (avatarPath) return AvatarImage({ large, avatarPath })
-  const codepoint = displayName && displayName.codePointAt(0)
-  const initial = codepoint
-    ? String.fromCodePoint(codepoint).toUpperCase()
-    : '#'
-
-  return (
-    <div
-      className={classNames('AvatarBubble', { large })}
-      style={{ backgroundColor: color }}
-    >
-      {initial}
-    </div>
-  )
-}
-
-export function AvatarImage({
-  avatarPath,
-  large,
-  ...otherProps
-}: {
-  avatarPath: string
-  large?: boolean
-} & React.ImgHTMLAttributes<HTMLImageElement>) {
-  return (
-    <img
-      className={classNames('AvatarImage', { large })}
-      src={avatarPath}
-      {...otherProps}
-    />
-  )
-}
-
-export function QRAvatar() {
-  return (
-    <div className='AvatarBubble'>
-      <img
-        className='avatar-qr-code-img sharp-pixel-image'
-        src='../images/qr_icon.png'
-      />
-    </div>
-  )
 }
 
 export const VerifiedIcon = (props: { style?: CSSProperties }) => (
@@ -129,39 +68,17 @@ export function PseudoContact(
   const { cutoff, text, subText } = props
   return (
     <div className='contact'>
-      {props.children ? props.children : renderAvatar(null, '#505050', cutoff)}
+      {props.children ? (
+        props.children
+      ) : (
+        <Avatar avatarPath={null} color={'#505050'} displayName={cutoff} />
+      )}
       {!subText && (
         <div className='contact-name'>
           <div className='pseudo-contact-text'>{text}</div>
         </div>
       )}
       {subText && ContactName(text, subText, false)}
-    </div>
-  )
-}
-
-export function AvatarBubble(
-  props: PropsWithChildren<
-    {
-      className?: string
-      noSearchResults?: boolean
-      large?: boolean
-    } & React.HTMLAttributes<HTMLDivElement>
-  >
-) {
-  return (
-    <div
-      className={classNames(
-        'AvatarBubble',
-        {
-          'AvatarBubble--NoSearchResults': props.noSearchResults,
-          large: props.large,
-        },
-        props.className
-      )}
-      {...props}
-    >
-      {props.children}
     </div>
   )
 }