+++ src/renderer/components/dialogs/ViewProfile.tsx
@@ -6,7 +6,6 @@ import {
   DeltaDialogContent,
   DeltaDialogContentTextSeperator,
 } from './DeltaDialog'
-import { Avatar } from '../contact/Contact'
 import { integerToHexColor } from '../../../shared/util'
 import ChatListItem from '../chat/ChatListItem'
 import { useChatListIds, useLazyChatListItems } from '../chat/ChatListHelpers'
@@ -16,6 +15,7 @@ import { Button } from '@blueprintjs/core'
 import { JsonContact } from '../../../shared/shared-types'
 import { C } from 'deltachat-node/dist/constants'
 import { ScreenContext } from '../../contexts'
+import { Avatar } from '../Avatar'
 
 const ProfileInfoName = ({ contactId }: { contactId: number }) => {
   const [contact, setContact] = useState<{